+++ cpukit/score/src/pheapallocatealigned.c
@@ -1,4 +1,12 @@
 /**
+ * @file
+ *
+ * @ingroup ScoreProtHeap
+ *
+ * @brief Protected Heap Handler implementation.
+ */
+
+/*
  *  COPYRIGHT (c) 1989-2007.
  *  On-Line Applications Research Corporation (OAR).
  *
@@ -18,8 +26,8 @@
 
 void *_Protected_heap_Allocate_aligned(
   Heap_Control *the_heap,
-  intptr_t      size,
-  uint32_t      alignment
+  uintptr_t     size,
+  uintptr_t     alignment
 )
 {
   void *p;
